fix: Modified the "Increment" button code so that the slider value is capped at 10.0.
The original code had a bug in it where if you clicked the increment button enough it
would increase the value beyond the maximum that the slider was able to represent.
This commit makes a modification so that the maximum value that the slider can ever
achieve is 10.0 (same as the original code), regardless of how many times you click
the increment button.  In addition, once you reach the maximum the button is disabled,
which makes it obvious that you can't go past 10.0.