add nix flake
this streamlines the UX for nix users, by managing dependencies and the
build process declaratively.

To build the template, a call to `nix build` suffices.
To run the template, a call to `nix run` suffics.
To develop the template, a call to `nix develop` will launch a shell
with all required dependencies in view. All of the existing scripts
should work from within this shell.